chore(ai): restructure agent files
