+++ .changeset/forty-apricots-hug.md
@@ -1,0 +1,5 @@
+---
+'ai': patch
+---
+
+chore(ai): restructure agent files
