+++ packages/ai/src/agent/agent-on-step-finish-callback.ts
@@ -1,0 +1,11 @@
+import { StepResult } from '../generate-text/step-result';
+import { ToolSet } from '../generate-text/tool-set';
+
+/**
+Callback that is set using the `onStepFinish` option.
+
+@param stepResult - The result of the step.
+ */
+export type AgentOnStepFinishCallback<TOOLS extends ToolSet> = (
+  stepResult: StepResult<TOOLS>,
+) => Promise<void> | void;
