+++ packages/ai/src/agent/index.ts
@@ -1,20 +1,24 @@
 export {
   Agent,
-  type AgentSettings,
-  type InferAgentUIMessage,
 
   /**
    * @deprecated Use `Agent` instead.
    */
   Agent as Experimental_Agent,
+} from './agent';
+export { type AgentOnStepFinishCallback } from './agent-on-step-finish-callback';
+export {
+  type AgentSettings,
 
   /**
    * @deprecated Use `AgentSettings` instead.
    */
   type AgentSettings as Experimental_AgentSettings,
-
+} from './agent-settings';
+export {
   /**
    * @deprecated Use `InferAgentUIMessage` instead.
    */
   type InferAgentUIMessage as Experimental_InferAgentUIMessage,
-} from './agent';
+  type InferAgentUIMessage,
+} from './infer-agent-ui-message';
