+++ packages/ai/src/agent/inter-agent-tools.ts
@@ -1,0 +1,7 @@
+import { Agent } from './agent';
+
+/**
+ * Infer the type of the tools of an agent.
+ */
+export type InferAgentTools<AGENT> =
+  AGENT extends Agent<infer TOOLS, any, any> ? TOOLS : never;
